feat: Set `DOCKER_HOST` var
Docker components are fine with the default socket path ($PREFIX/var/run/docker.sock), but I changed the Docker path to /data/docker (Which is better IMO).
That change made some things to broke, and with the var `DOCKER_HOST` I fixed it.
No need to have a wrapper specifically for Docker Composer.